feat: Add activity page and controller
This commit adds a new route, controller, and view for the activities page. The route is defined in the web.php file, and the controller is created as activity_Controller.php. The view file is added in the resources/views folder. This allows users to access the activities page and view the available activities.

#changelog
- Added activity page and controller